MultiWiki functionality is now an extension to JSPWiki in a separate module
* added module jspwiki-multiwiki
* updated documentation how to activate extensions
* more work on tests
* minor changes to WikiEngine initialization to allow for extensions by PageRenamer and PageNameResolver

(cherry picked from commit faf98c3d2b9ef1d638ef414c6eee4f898d1fa9a3)
