+++ jspwiki-main/src/main/java/org/apache/wiki/WikiPage.java
@@ -18,6 +18,7 @@
  */
 package org.apache.wiki;
 
+
 import org.apache.wiki.api.core.Engine;
 import org.apache.wiki.api.core.Page;
 import org.apache.wiki.api.providers.PageProvider;
@@ -25,7 +26,7 @@ import org.apache.wiki.auth.acl.Acl;
 import org.apache.wiki.auth.acl.AclEntry;
 import org.apache.wiki.auth.acl.AclImpl;
 import org.apache.wiki.pages.PageManager;
-import org.apache.wiki.providers.SubWikiUtils;
+import org.apache.wiki.pages.PageNameResolverManager;
 
 import java.util.Date;
 import java.util.Enumeration;
@@ -59,7 +60,8 @@ public class WikiPage implements Page {
      */
     public WikiPage( final Engine engine, final String name ) {
         m_engine = engine;
-        m_name = SubWikiUtils.expandPageNameWithMainPrefix(name, m_engine.getWikiProperties());
+        PageNameResolverManager manager = engine.getManager(PageNameResolverManager.class);
+        m_name = manager.getPageNameResolver().resolvePageName(name, m_engine.getWikiProperties());
         m_wiki = engine.getApplicationName();
     }
