+++ jspwiki-main/src/main/java/org/apache/wiki/pages/PageNameResolverManager.java
@@ -1,0 +1,14 @@
+/*
+ * Copyright (C) 2025 denkbares GmbH. All rights reserved.
+ */
+
+package org.apache.wiki.pages;
+
+import org.jetbrains.annotations.NotNull;
+
+@FunctionalInterface
+public interface PageNameResolverManager {
+
+	@NotNull PageNameResolver getPageNameResolver();
+
+}
