+++ jspwiki-multiwiki/src/main/java/org/apache/wiki/content/DefaultPageRenamerMultiWiki.java
